Move components to different folders
